Fryer refactor WIP
